+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add `to_nanos()`, `to_micros()`, `to_millis()`, `to_secs()`, `to_minutes()`,
   and `to_hours()` methods to `Duration<u32, NOM, DENOM>` and
   `Duration<u64, NOM, DENOM>` types to easily convert to integer time units.
+- Support for `Rate` (Hertz, Kilohertz, ...)
 
 ### Fixed